SceneMgr drived from PGWorld
