+++ pgdrive/engine/world/image_buffer.py
