+++ pgdrive/engine/world/manual_controller.py
@@ -1,7 +1,7 @@
 from direct.controls.InputState import InputState
 
 from pgdrive.utils import import_pygame
-from pgdrive.pgdrive_engine.world.pg_world import PGWorld
+from pgdrive.engine.world.pg_world import PGWorld
 
 pygame = import_pygame()
