--- pgdrive/pgdrive_engine/__init__.py
@@ -1,1 +1,0 @@
-from pgdrive.pgdrive_engine.pgdrive_engine import initialize_pgdrive_engine
