+++ pgdrive/scene_creator/vehicle_module/mini_map.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 from panda3d.core import Vec3, NodePath
 from pgdrive.constants import CamMask
-from pgdrive.pgdrive_engine.world.image_buffer import ImageBuffer
+from pgdrive.engine.world.image_buffer import ImageBuffer
 
 
 class MiniMap(ImageBuffer):
