+++ pgdrive/scene_creator/vehicle_module/rgb_camera.py
@@ -1,8 +1,8 @@
 from panda3d.core import Vec3, NodePath
 
 from pgdrive.constants import CamMask
-from pgdrive.pgdrive_engine.world.image_buffer import ImageBuffer
-from pgdrive.pgdrive_engine.world.pg_world import PGWorld
+from pgdrive.engine.world.image_buffer import ImageBuffer
+from pgdrive.engine.world.pg_world import PGWorld
 
 
 class RGBCamera(ImageBuffer):
