+++ pgdrive/tests/test_env/test_build_city.py
@@ -1,6 +1,6 @@
 from pgdrive import PGDriveEnv
 from pgdrive.scene_creator.city_map import CityMap
-from pgdrive.pgdrive_engine.world.pg_world import PGWorld
+from pgdrive.engine.world.pg_world import PGWorld
 
 
 def _t(num_blocks):
